Add preruns to create schemas before processing any data
To avoid a race condition on schemas, do pre-run for each task
with no input science data, and pre-generate the schemas. Treat
schemas as the input files of actual runs with input science data;
track schema files with Pegasus.

Do not pre-run makeSkyMap.py as it would do the actual run.
For now, do not pre-run makeCoaddTempExp and assembleCoadd as
they do not use schemas.

For now, ignore config/versions although they are subject to
similar problem of race condition.