Make task adding more uniform
Names of local variables were tightly coupled with tasks they were
related to. It makes hard to factor out some common patterns in the
process of adding a task despite obvious similarities.  To prepare the
ground for factoring out this process to a separate function I renamed
many variables in a more generic fashion focusing on their functional
role rather than emphasizing the task the belong to.