Add blacklistCoaddTempExp to exclude some patch-visit pairs
In makeCoaddTempExp, these patch-visit pairs had 0 good pixels
generated in their directWarp. Example log message:

  makeCoaddTempExp INFO: Processing calexp 1 of 1 for this Warp: id={...}
  makeCoaddTempExp INFO: directWarp has 0 good pixels (0.0%)
  makeCoaddTempExp WARN: Warp DataId(initialdata={'filter': 'HSC-I', 'tract': 8766, 'visit': 7300, 'patch': '3,7'}, tag=set([])) could not be created

This is troublesome because no wrap is made although we only know
if after running it. In the future a dummy placeholder file will
be created so these can run as usual.

Here I use a blacklist of tract-patch-visit so to ignore them
in makeCoaddTempExp.